pyproject.toml
[tool.poetry]
name = "satellitevu"
version = "5.2.0"
description = "Client SDK for SatelliteVu's platform APIs"
authors = ["Christian Wygoda <christian.wygoda@satellitevu.com>", "Zhelini Sivanesan <zhelini.sivanesan@satellitevu.com>", "James Harrison <james.harrison@satellitevu.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
appdirs = "^1.4.4"
allure-pytest = "^2.13.5"

[tool.poetry.group.dev.dependencies]
allure-pytest = "^2.13.5"
cryptography = "^43.0.3"
fawltydeps = "^0.17.0"
josepy = "^1.14.0"
mocket = "^3.13.2"
nox = "^2024.10.9"
nox-poetry = "^1.0.1"
pyfakefs = "^5.7.1"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pyjwt = "^2.9.0"
ruff = "^0.7.3"
pact-python = "^2.2.2"

[tool.poetry.group.requests]
optional = true

[tool.poetry.group.requests.dependencies]
requests = "^2.28.1"

[tool.poetry.group.notebooks]
optional = true

[tool.poetry.group.notebooks.dependencies]
folium = "^0.13.0"
geopandas = "^0.12.1"
pandas = "^1.5.1"
shapely = "^1.8.5.post1"

[tool.poetry.group.httpx]
optional = true

[tool.poetry.group.httpx.dependencies]
httpx = "^0.23.0"

[tool.commitizen]
tag_format = "v$version"
version_scheme = "semver2"
version_provider = "poetry"
update_changelog_on_bump = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.fawltydeps]
ignore_unused = [
    "fawltydeps",
    "nox",
    "pytest-cov",
    "ruff",
    "pact-python",
]

[tool.fawltydeps.custom_mapping]
allure-pytest = ["allure"]
pyjwt = ["jwt"]
pact-python = ["pact"]
